% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_N2000_monitoring}
\alias{get_N2000_monitoring}
\title{Returns a dataframe with monitoring requirements per Natura2000 habitattype}
\usage{
get_N2000_monitoring(habitattype = "all")
}
\arguments{
\item{habitattype}{character with habitattype code,'all' returns all beheertypen. See https://www.natura2000.nl/beschermde-natuur/habitattypen for details.}
}
\description{
Returns a dataframe with monitoring requirements per Natura2000 habitattype
}
\examples{
library(VHRscope)

get_SNL_monitoring_overview()

}
