% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_soortenlijst}
\alias{get_soortenlijst}
\title{Returns a dataframe with species and metadata from NDFF and SOVON}
\usage{
get_soortenlijst()
}
\description{
Returns a dataframe with species and metadata from NDFF and SOVON
}
\examples{
library(VHRscope)

get_soortenlijst()

}
