% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{habitathoofdtype}
\alias{habitathoofdtype}
\title{Returns 'hoofdtype' van habitattypecode}
\usage{
habitathoofdtype(habitattype)
}
\arguments{
\item{habitattype}{character vector with habittattype codes}
}
\description{
Returns 'hoofdtype' van habitattypecode
}
\examples{
example code


library(VHRscope)

habitathoofdtype('H2180A')
}
