% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{habitattype}
\alias{habitattype}
\title{Returns 'habitattype' van habitatsubtypecode}
\usage{
habitattype(habitatsubtype)
}
\arguments{
\item{habitattype}{character vector with habittattype codes}
}
\description{
Returns 'habitattype' van habitatsubtypecode
}
\examples{
library(VHRscope)

habitattype(habitatsubtype = 'H2190Aom')

}
