% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scale_fill_habitats}
\alias{scale_fill_habitats}
\title{Create discrete scale for Natura 2000 habitat types}
\usage{
scale_fill_habitats(palette = "PZ", legend = "code", ...)
}
\arguments{
\item{palette}{character to indicate color palette: default 'PZ'.}

\item{legend}{character to indicate the legend appearance: 'code', 'name' or 'name_code'}
}
\description{
ggplot2 helper function to create a discrete scale for Natura 2000 habitat types
}
