% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{st_full_union}
\alias{st_full_union}
\title{Performs a full union}
\usage{
st_full_union(x, y)
}
\arguments{
\item{x}{An sf object.}

\item{y}{An sf object.}
}
\value{
An sf object.
}
\description{
Performs a full union
}
\examples{
library(VHRscope)
library(sf)
library(dplyr)
library(ggplot2)

pol1 = st_polygon(list(rbind(c(0,0), c(2,0), c(2,2), c(0,2), c(0,0))))
pol2 = st_polygon(list(rbind(c(2,0), c(4,0), c(4,2), c(2,2), c(2,0))))

A <- st_as_sf(st_sfc(pol1,pol2)) \%>\% mutate(cat = "A")

pol3 = st_polygon(list(rbind(c(1,1), c(3,1), c(3,3), c(1,3), c(1,1))))
pol4 = st_polygon(list(rbind(c(3,1), c(5,1), c(5,3), c(3,3), c(3,1))))

B <- st_as_sf(st_sfc(pol3,pol4)) \%>\% mutate(cat = "B")

ggplot(data = A)+
 geom_sf(fill = 'red', alpha = .5)+
 geom_sf(data = B, fill = 'blue', alpha = .5)

C <- st_full_union(x = A, y = B)

ggplot(data = C)+
 geom_sf(aes(fill = cat.x, color = cat.y), alpha = .5)


}
