% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{st_read_kmhokken}
\alias{st_read_kmhokken}
\title{Returns an sf object with Dutch (NDFF) kilometer-hokken}
\usage{
st_read_kmhokken()
}
\description{
Returns an sf object with Dutch (NDFF) kilometer-hokken
}
\examples{
library(VHRscope)

sf_km <- st_read_kmhokken()


}
