% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{st_read_N2000NL}
\alias{st_read_N2000NL}
\title{Returns an sf object with Dutch N2000 areas}
\usage{
st_read_N2000NL(gebiedNummer = "all")
}
\arguments{
\item{gebiedNummer}{character with official number of the area,'all' returns all areas. See https://www.natura2000.nl/gebieden for details.}
}
\description{
Returns an sf object with Dutch N2000 areas
}
\examples{
library(VHRscope)

sf_N2000 <- st_read_N2000NL()

plot((sf_N2000 \%>\% filter(gebiedNummer \%in\% 116:126))[,1])

}
